snowflake_agent/requirements.txt
snowflake-snowpark-python>=1.0.0
snowflake-connector-python>=3.0.0
pandas>=1.5.0
numpy>=1.23.0
cryptography>=3.4.0
